Add a A::Web#local_redirect method
(Suggestions for a better/shorter name are welcome)

This one acts just like redirect() with the only difference that it
strips the $url_scheme://$HTTP_HOST at the beginning of the url.

Use case is as follows :
 - Say your application is behind a (dumb!) proxy that handles the
 requests on port 80.
 - Your app runs on standard Starlet (0:5000).
 - You call redirect('/foo')
 - the proxy returns the header : Location: http://0:5000/foo
 - DOKAN!

With this new method, the returned header is : "Location: /foo"

If the app runs under a SCRIPT_NAME '/my_app/' then the header will be :
"Location: /my_app/foo".